Refactor: 패키징 리팩토링
