+++ app/src/main/java/choijjyo/reco/Like/Fragment_Color.kt
@@ -1,4 +1,4 @@
-package choijjyo.reco
+package choijjyo.reco.Like
 
 import android.graphics.Color
 import android.os.Bundle
